Add ability to selectively disable coverage checks.
For most types and properties, we want to verify that we have covered
various conditions and fail the property test if we are not able to
cover all conditions.

But for some types, such as `()`, certain conditions can never be true.
For example, there can never be a value of type `()` that is not equal
to `mempty`.

To make it possible to selectively disable coverage checks, we define
the `disableCoverageCheck` function.