Miscellaneous improvements to `internal` library.
The `internal` library within `quickcheck-groups` was originally a fork
of the `internal` library within `quickcheck-monoid-subclasses`.

This commit brings the `internal` library up to date with recent changes
made to the `quickcheck-monoid-subclasses' library.

In future, we should consider extracting out this `internal` library to
a separate library on hackage, to avoid code duplication.